scripts: rework ntp interface address updates
This is rewrite of the code handling update and deletion
of addresses to update ntp.conf

The original code allowed one address to configured in
'interface listen' configuration in ntp.conf. Since
ntp.conf may have both IPV4 and IPV6 addresses, the code
needs to handle multiple listen addresses.

This commit also handles a condition with multiple IPV4 or
IPV6 addresses on an interface where the correct behavior
on deleting an IP address should be using another address
from the interface instead of stop listening to these addresses.

VRVDR-51135